ethdev: introduce device reinitialisation API
Add a new ethdev API for resetting a device followed by a restoration of
the configuration present at the time of the invocation of the API.
In some cases if the driver can support it, the device will be restarted
if it was started when the API was invoked.

This operation could be useful for example when the user has detected an
issue with the device and wants to reset and reinitialise it in an
attempt to get it working again. While this is possible by calling a
sequence of existing APIs (reset, configure, rx_queue_setup and
tx_queue_setup), in the case where we are restoring the exact same
configuration, much of the reconfiguration performed in those calls
would be unnecessary and can be achieved by calling a subset of that
functionality, which is what this API will perform.

Signed-off-by: Ciara Loftus <ciara.loftus@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>