bus/cdx: fix incorrect device release on error
The variable cdx_addr is a stack-allocated zero-initialized char array, but
it is never used anywhere except in an error handler to stand it as a
device bus address for a VFIO device release call. This is incorrect, as it
should have been actual device address, so fix the code to reflect that.

Fixes: 45ef232af515 ("bus/cdx: introduce AMD CDX bus")
Cc: nipun.gupta@amd.com
Cc: stable@dpdk.org

Signed-off-by: Anatoly Burakov <anatoly.burakov@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>