dts: add method for bringing link of interface up
Add set_interface_link_up method to Linux Session class
for setting the link of an interface up. The existing
command most similar to this, bring_up_link, is designed
for use during topology setup within the framework,
whereas the new command is designed for use during
test suite runtime, as is needed for the virtio suite.

Signed-off-by: Dean Marx <dmarx@iol.unh.edu>
Signed-off-by: 0-day Robot <robot@bytheb.org>