net/tap: configure link carrier
In order for net_tap interfaces to be used in linux bonds, they need to
report a carrier.

After at last one RX queue has been configured, force the carrier on
with ioctl(TUNSETCARRIER) (there is no netlink equivalent for this).

Also do the same after the device has been moved to another netns.

TUNSETCARRIER support was added in Linux 5.0. Only enable carrier set if
this ioctl number is defined.

Signed-off-by: Robin Jarry <rjarry@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>