test/memcpy: reduce duration
When running on limited platforms like GitHub Actions,
the functional unit test "memcpy_autotest"
may hit a timeout, especially when running with UBSan.

This change skips testing some alignment offsets (from 0 to 31).
It will test only every 3-byte offsets for the source buffer,
and every 2-byte offsets for the destination buffer.
So it is supposed to be 6x faster with a reasonably smaller coverage.

Signed-off-by: Thomas Monjalon <thomas@monjalon.net>
Acked-by: Morten Brørup <mb@smartsharesystems.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>