vhost_user: Increase number of memory regions
In this patch the number of memory regions are increased from
8 to 128. When a vhost-user front-end such as qemu or libblkio
queries the back-end such as dpdk with the message, get max number
of memory slots, the back-end replies with this number 128 instead
of the previously defined 8. The back-end also allocates that many
slots in the memory table where regions are added/removed as
requested by the vhost-user front-end. This also helps the vhost-
user front-end to limit the number of memory regions when sending
the set mem table message ar adding memory regions.

Signed-off-by: Pravin M Bathija <pravin.bathija@dell.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>