net/mlx5: rework root group checks in table create
Before this patch, during flow actions translation done in template
table create, there were a lot of checks for group with index 0.
This group is special, because flow rules in that group are created
through mlx5 kernel driver.
For this reason, this group is referred to as root group or root table.

This patch reworks these group index checks for clarity:

- A dedicated function for checking if group index refers to
  root group is added.
- All direct group index checks in actions translation are replaced
  with a check for result of that function.
- Redundant group translation checks in actions translation is removed.
  Group indexes are already translated at that point.
- Root group check for internal default miss action is removed.
  This action is supported on root group assuming rdma-core supports it.
  If rdma-core core does not support it, then HWS layer will reject
  accordingly.

Signed-off-by: Dariusz Sosnowski <dsosnowski@nvidia.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>