crypto/qat: fix CCM request descriptor hash_state_sz field
The hash_state_sz field in the QAT request descriptor for AES-CCM
operations was incorrectly set to the digest length. According to
the QAT firmware specification, this field should contain the
length of AAD in quadwords for CCM operations.

This incorrect field value caused QAT firmware to return
INVALID_PARAMS errors for certain AAD buffer sizes. A new unit
test is also added to cover this case.

Fixes: ab56c4d ("crypto/qat: add AES-CCM support")

Signed-off-by: Emma Finn <emma.finn@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>