vhost: fix use-after-free race during cleanup
This commit fixes a use-after-free that causes the application to crash
on shutdown (detected by ASAN).

The vhost library uses a background event dispatch thread that monitors
fds with epoll. It runs in an infinite loop, waiting for I/O events
and calling callbacks when they occur.

During cleanup, a race condition existed:

  Main Thread:                    Event Dispatch Thread:
  1. Remove fds from fdset        while (1) {
  2. Close file descriptors           epoll_wait() [gets interrupted]
  3. Free fdset memory                [continues loop]
  4. Continue...                      Accesses fdset...   CRASH
                                  }

There isn't explicit cleanup of the fdset structure.
The fdset structue is allocated with rte_zmalloc() and the memory would
only be reclaimed at application shutdown when rte_eal_cleanup() is called,
which invokes rte_eal_memory_detach() to unmap all the hugepage memory.
Meanwhile, the event dispatch thread could still be running and accessing
the fdset.

The code had a `destroy` flag that the event dispatch thread checked,
but it was never set during cleanup, and the code never waited for
the thread to actually exit before freeing memory.

To fix this, the commit implements `fdset_destroy()` that will set the
destroy flag, wait for thread termination, and clean up all resources.

The socket.c and vduse.c are updated to call fdset_destroy() when
the last socket/device is unregistered.

For vduse, reference counting was added to track the number of devices
using the fdset. The fdset is destroyed when the last device is removed.

Fixes: 0e38b42bf61c ("vhost: manage FD with epoll")
Cc: stable@dpdk.org

Signed-off-by: Shani Peretz <shperetz@nvidia.com>

----------
v2:
- Extended the fix to vduse.c, added reference counting and mutex to vduse
  structure to track the number of devices using the fdset
- Call fdset_destroy() when last device is removed in vduse_device_destroy()
  and in error paths of vduse_device_create()
- Added mutex protection when checking/setting the destroy flag to prevent
  race conditions in both fdset_event_dispatch() and fdset_destroy()

Signed-off-by: 0-day Robot <robot@bytheb.org>