dts: add performance test functions and support
Provide packet transmission function to support performance tests using a
user-supplied performance traffic generator. The single core performance
test is included. It allows the user to define a matrix of frame size,
descriptor count, and expected mpps, and fails if any combination does
not forward a mpps count within 5% of the given baseline. Furthermore
ensure that the DPDK build on the SUT is including the correct arguments
to allow for the highest possible packet throughput.

Bugzilla ID: 1697
Signed-off-by: Nicholas Pratte <npratte@iol.unh.edu>
Signed-off-by: Patrick Robb <probb@iol.unh.edu>
Reviewed-by: Dean Marx <dmarx@iol.unh.edu>
Reviewed-by: Andrew Bailey <abailey@iol.unh.edu>
Signed-off-by: 0-day Robot <robot@bytheb.org>