net/crc: remove redundant operations in crcr32_reduce_64_to_32
Replace the clearing of lower 32 bits of XMM register with blend of
zero register.
Remove the clearing of upper 64 bits of tmp1 as it is redundant.
tmp1 after clearing upper bits was being xor with tmp2 before the
bits 96:65 from tmp2 were returned. The xor operation of bits 96:65
remains unchanged due to tmp1 having bits 96:64 cleared to 0.
After removing the xor operation, the clearing of upper 64 bits of tmp1
becomes redundant and hence can be removed.

Signed-off-by: Shreesh Adiga <16567adigashreesh@gmail.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>