net/mlx5: fix the masked value for a rule for sync API
When inserting a rule via HWS synchronous API, the underlayer
implementaiton is a bit different from HWS. The template(async)
API calls are re-used. In the template API definition, the user
should ensure that there is no value bit that is not masked by
the template.

In the legacy SWS synchronous API, when translating the item, if a
mask is provided together with the spec. The value is the result of
the spec OP-AND mask. If no mask, a default mask with all 1s are used
and the value of the spec remain the original input.

By introducing the new logic to do the OP-AND and duplicate the
items from the input in the rte_flow layer. The spec field will
be with the correct value after inline OP-AND calculating. The rule
will be inserted with proper value as expected by the user.

Fixes: e38776c36c8a ("net/mlx5: introduce HWS for non-template flow API")
Cc: mkashani@nvidia.com

Signed-off-by: Bing Zhao <bingz@nvidia.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>