net/txgbe: fix the missing old mailbox interface calls
For different devices, the mailbox flow between software and firmware is
different. There are several mailbox command in the txgbe driver, but only
txgbe_hic_sr_read() was changed to use the new flow.

It lead to other mailbox commands timeout for Amber-Lite devices, which
is required to use the new flow. So this patch fills in the missing part.

For the sake of code tidy, txgbe_hic_sr_read() should change to use the
locked function txgbe_host_interface_command(), and this function could
be change to pointer in struct txgbe_mbx_info for different devices.

Fixes: 6a139ade82e7 ("net/txgbe: add new SW-FW mailbox interface")
Signed-off-by: Jiawen Wu <jiawenwu@trustnetic.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>